Handle dynamic category assignment in constructor
