+++ packages/framework/src/Models/Pages/DocumentationPage.php
@@ -6,6 +6,7 @@ use Hyde\Framework\Concerns\HasTableOfContents;
 use Hyde\Framework\Contracts\AbstractMarkdownPage;
 use Hyde\Framework\Contracts\RouteContract;
 use Hyde\Framework\Models\Route;
+use Illuminate\Support\Str;
 
 class DocumentationPage extends AbstractMarkdownPage
 {
@@ -20,10 +21,19 @@ class DocumentationPage extends AbstractMarkdownPage
      */
     public ?string $category;
 
-    public function __construct(array $matter = [], string $body = '', string $title = '', string $slug = '', ?string $category = null)
+    public function __construct(array $matter = [], string $body = '', string $title = '', string $slug = '')
     {
         parent::__construct($matter, $body, $title, $slug);
-        $this->category = $category;
+        $this->category = $this->getDocumentationPageCategory();
+    }
+
+    protected function getDocumentationPageCategory(): ?string
+    {
+        if (str_contains($this->slug, '/')) {
+            return Str::before($this->slug, '/');
+        }
+
+        return $this->matter['category'] ?? null;
     }
 
     /** @inheritDoc */
