+++ packages/framework/tests/Feature/SourceFileParserTest.php
@@ -60,30 +60,4 @@ class SourceFileParserTest extends TestCase
         $this->assertEquals('# Foo Bar', $page->body);
         $this->assertEquals('Foo Bar Baz', $page->title);
     }
-
-    public function test_documentation_page_parser_can_get_category_from_front_matter()
-    {
-        $this->markdown('_docs/foo.md', '# Foo Bar', ['category' => 'foo']);
-
-        $parser = new SourceFileParser(DocumentationPage::class, 'foo');
-
-        /** @var DocumentationPage $page */
-        $page = $parser->get();
-        $this->assertEquals('foo', $page->category);
-    }
-
-    public function test_documentation_page_parser_can_get_category_automatically_from_nested_page()
-    {
-        mkdir(Hyde::path('_docs/foo'));
-        touch(Hyde::path('_docs/foo/bar.md'));
-
-        $parser = new SourceFileParser(DocumentationPage::class, 'foo/bar');
-
-        /** @var DocumentationPage $page */
-        $page = $parser->get();
-        $this->assertEquals('foo', $page->category);
-
-        unlink(Hyde::path('_docs/foo/bar.md'));
-        rmdir(Hyde::path('_docs/foo'));
-    }
 }
