+++ packages/framework/tests/Unit/DocumentationPageTest.php
@@ -2,6 +2,7 @@
 
 namespace Hyde\Framework\Testing\Unit;
 
+use Hyde\Framework\Actions\SourceFileParser;
 use Hyde\Framework\Hyde;
 use Hyde\Framework\HydeServiceProvider;
 use Hyde\Framework\Models\Pages\DocumentationPage;
@@ -127,5 +128,31 @@ class DocumentationPageTest extends TestCase
     {
         $page = (new DocumentationPage([], '', '', 'foo/bar'));
         $this->assertEquals('docs/bar.html', $page->getOutputPath());
+    }
+
+    public function test_documentation_page_parser_can_get_category_from_front_matter()
+    {
+        $this->markdown('_docs/foo.md', '# Foo Bar', ['category' => 'foo']);
+
+        $parser = new SourceFileParser(DocumentationPage::class, 'foo');
+
+        /** @var DocumentationPage $page */
+        $page = $parser->get();
+        $this->assertEquals('foo', $page->category);
+    }
+
+    public function test_documentation_page_parser_can_get_category_automatically_from_nested_page()
+    {
+        mkdir(Hyde::path('_docs/foo'));
+        touch(Hyde::path('_docs/foo/bar.md'));
+
+        $parser = new SourceFileParser(DocumentationPage::class, 'foo/bar');
+
+        /** @var DocumentationPage $page */
+        $page = $parser->get();
+        $this->assertEquals('foo', $page->category);
+
+        unlink(Hyde::path('_docs/foo/bar.md'));
+        rmdir(Hyde::path('_docs/foo'));
     }
 }
